\name{enzymes}
\docType{data}
\alias{enzymes}
\title{Liver Enzyme Data}
\description{Liver enzyme data collected from 218 patients with liver disease (Plomteux, 1980). The laboratory profile consists of enzymatic activity measured for four liver enzymes: aspartate aminotransferase (AST), alanine aminotransferase (ALT), glutamate dehydrogenase (GLDH) and ornithine carbamyltransferase (OCT).}
\usage{data(enzymes)}
\format{A data frame with 218 rows and the following 6 columns:
  \describe{
    \item{\code{Patient}}{Patient ID}
    \item{\code{Group}}{Four diagnostic groups were considered: acute viral hepatitis (1), persistent chronic hepatitis (2), aggressive chronic hepatitis (3) and post-necrotic cirrhosis (4).}
    \item{\code{AST}}{Aspartate aminotransferase (U/L)}
    \item{\code{ALT}}{Alanine aminotransferase (U/L)}
    \item{\code{GLDH}}{Glutamate dehydrogenase (U/L}
    \item{\code{OCT}}{Ornithine carbamyltransferase (U/L)}
  }
 } 
\source{Albert, A. and Harris, E. K. (1984) \emph{Multivariate Interpretation of Clinical Laboratory Data}. Dekker: New York, Chapter 5, Appendix I.}
\references{Plomteux, G. (1980) Multivariate analysis of an enzyme profile for the differential diagnosis of viral hepatitis. \emph{Clinical Chemistry}, \bold{26}, 1897--1899.}
\keyword{datasets}

