% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pmlr.R
\name{summary.pmlr}
\alias{summary.pmlr}
\title{Summarizing Penalized Multinomial Logistic Model Fits}
\usage{
\method{summary}{pmlr}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"pmlr"}, a result of a call to \code{\link{pmlr}}.}

\item{...}{further arguments passed to or from other methods}
}
\value{
\code{summary.pmlr} returns an object of class \code{"summary.pmlr"}, a list with components
\item{call}{the matched call of the \code{object}}
\item{method}{which method was used for hypothesis testing and computation of confidence intervals}
\item{coef}{an array containing the coefficient estimates, standard errors, and
test statistics and their p-values asssociated with the chosen method for the p parameters for the J categories}
\item{joint.test}{an array contatining the test statistics and p-values from constrained hypothesis tests under all betas = 0,
all betas are equal, and betas are proportional.}
\item{test.all0.vs.constraint}{Returned only if joint hypothesis testing was done:
An array containing likelihood ratio test statistics and p-values testing all \eqn{H_0}: betas=0 vs. other constraints (\eqn{H_C}),
which are 'all betas are equal' and 'betas are proportion'.}
}
\description{
This function is for class \code{pmlr} object.
}

